remove projector
